+++ app/lib/shared/versions.dart
@@ -24,10 +24,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2025.02.25',
+  '2025.03.03',
   // Fallback runtime versions.
+  '2025.02.25',
   '2025.02.14',
-  '2025.02.10',
 ];
 
 /// Sets the current runtime versions.
@@ -62,7 +62,7 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '3.7.0';
-final String toolStableDartSdkVersion = '3.7.0';
+final String toolStableDartSdkVersion = '3.7.1';
 final String toolStableFlutterSdkVersion = '3.29.0';
 
 final semanticToolStableDartSdkVersion =
@@ -74,4 +74,4 @@ final semanticToolStableFlutterSdkVersion =
 final String panaVersion = pana.packageVersion;
 
 // keep in-sync with pkg/pub-worker/lib/src/bin/pana_wrapper.dart
-final String dartdocVersion = '8.3.2';
+final String dartdocVersion = '8.3.3';