+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bump runtimeVersion to `2025.03.03`.
+ * Upgraded stable Dart analysis SDK to `3.7.1`
+ * Upgraded dartdoc to `8.3.3`.
 
 ## `20250225t105200-all`
  * Bump runtimeVersion to `2025.02.25`.